requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
croniter==2.0.5
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
schedule==1.2.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzlocal==5.2
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
